+++ platform/src/codatplatform/models/operations/configure_supplemental_data.py
@@ -3,12 +3,11 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import supplementaldataconfiguration as shared_supplementaldataconfiguration
+from ...models.shared import supplementaldataconfiguration as shared_supplementaldataconfiguration
 from enum import Enum
 from typing import Optional
 
-class ConfigureSupplementalDataDataType(str, Enum):
+class DataType(str, Enum):
     r"""Data types that support supplemental data"""
     CHART_OF_ACCOUNTS = 'chartOfAccounts'
     BILLS = 'bills'
@@ -35,7 +34,7 @@ class ConfigureSupplementalDataDataType(str, Enum):
 
 @dataclasses.dataclass
 class ConfigureSupplementalDataRequest:
-    data_type: ConfigureSupplementalDataDataType = dataclasses.field(metadata={'path_param': { 'field_name': 'dataType', 'style': 'simple', 'explode': False }})
+    data_type: DataType = dataclasses.field(metadata={'path_param': { 'field_name': 'dataType', 'style': 'simple', 'explode': False }})
     r"""Supported supplemental data data type."""
     platform_key: str = dataclasses.field(metadata={'path_param': { 'field_name': 'platformKey', 'style': 'simple', 'explode': False }})
     r"""A unique 4-letter key to represent a platform in each integration. View [accounting](https://docs.codat.io/integrations/accounting/overview#platform-keys), [banking](https://docs.codat.io/integrations/banking/overview#platform-keys), and [commerce](https://docs.codat.io/integrations/commerce/overview#platform-keys) platform keys."""
@@ -51,8 +50,6 @@ class ConfigureSupplementalDataResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your API request was not properly authorized."""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
     r"""Raw HTTP response; suitable for custom response parsing"""