+++ platform/src/codatplatform/models/operations/list_integrations.py
@@ -3,8 +3,7 @@
 from __future__ import annotations
 import dataclasses
 import requests as requests_http
-from ..shared import errormessage as shared_errormessage
-from ..shared import integrations as shared_integrations
+from ...models.shared import integrations as shared_integrations
 from typing import Optional
 
 
@@ -28,8 +27,6 @@ class ListIntegrationsResponse:
     r"""HTTP response content type for this operation"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    error_message: Optional[shared_errormessage.ErrorMessage] = dataclasses.field(default=None)
-    r"""Your `query` parameter was not correctly formed"""
     integrations: Optional[shared_integrations.Integrations] = dataclasses.field(default=None)
     r"""OK"""
     raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)